+++ maint/users.sql
@@ -1,24 +1,6 @@
 /*
- * When ACLs are changed related to pending certified products, remove this function
+ * create contacts
  */
-create or replace function openchpl.insert_sid_for_pending_cp (
-        sid bigint
-        )
-    returns void as $$
-    declare
-    rec record;
-    get_object text;
-
-    begin
-    get_object := 'select id from openchpl.acl_object_identity where object_id_class = 3';
-
-    for rec in execute get_object
-    loop
-    execute 'insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure) values ($1, (select max(ace_order) + 1 from openchpl.acl_entry where acl_object_identity = $1), $2, 16, true, false, false);' using rec.id, sid;
-    end loop;
-    end;
-    $$ language plpgsql;
-
 insert into openchpl.contact (first_name, last_name, email, phone_number, signature_date, last_modified_user) values
     ('Andrew', 'Larned', 'alarned@ainq.com', '301-560-6999', now(), -2)
     ,('Andrew', 'ACB', 'alarned@ainq.com', '301-560-6999', now(), -2)
@@ -27,6 +9,9 @@ insert into openchpl.contact (first_name, last_name, email, phone_number, signat
     ,('Ashwini', 'More','amore@ainq.com', '301-560-6999', now(), -2)
     ;
 
+/*
+ * create users
+ */
 insert into openchpl.user (user_name, password, account_expired, account_locked, credentials_expired, account_enabled, compliance_signature, last_modified_user, contact_id) values
     ('andlar-test', '$2a$10$D039AXXR3qZAoDtotSCj.OlysLVNy8/K6l4Jup4AuOdRrQ1haLwnq', false, false, false, true, now(), -2, (select contact_id from openchpl.contact where first_name = 'Andrew' and last_name = 'Larned' limit 1))
     ,('andlar-test-acb', '$2a$10$D039AXXR3qZAoDtotSCj.OlysLVNy8/K6l4Jup4AuOdRrQ1haLwnq', false, false, false, true, now(), -2, (select contact_id from openchpl.contact where first_name = 'Andrew' and last_name = 'ACB' limit 1))
@@ -35,6 +20,9 @@ insert into openchpl.user (user_name, password, account_expired, account_locked,
     ,('amore-icsa', '$2a$10$Xuc4/EDtgYr9bkcfUSxN6eUN8l/S796FZZL0k/PlJLMFtoa5hXM5i', false, false, false, true, now(), -2, (select contact_id from openchpl.contact where first_name = 'Ashwini' and last_name = 'More' limit 1))
     ;
 
+/*
+ * give users permission on themselves
+ */
 insert into openchpl.global_user_permission_map (user_id, user_permission_id_user_permission, last_modified_user) values
     ((select user_id from openchpl.user where user_name = 'andlar-test'), -2, -2)
     ,((select user_id from openchpl.user where user_name = 'andlar-test-acb'), 2, -2)
@@ -43,6 +31,9 @@ insert into openchpl.global_user_permission_map (user_id, user_permission_id_use
     ,((select user_id from openchpl.user where user_name = 'amore-icsa'), 2, -2)
     ;
 
+/*
+ * create sids
+ */
 insert into openchpl.acl_sid (principal, sid) values
     (true, 'andlar-test')
     ,(true, 'andlar-test-acb')
@@ -51,6 +42,9 @@ insert into openchpl.acl_sid (principal, sid) values
     ,(true, 'amore-icsa')
     ;
 
+/*
+ * identify which sid belongs to which user
+ */
 insert into openchpl.acl_object_identity (object_id_class, object_id_identity, parent_object, owner_sid, entries_inheriting) values
     (1, (select user_id from openchpl.user where user_name = 'andlar-test'), null, -2, true)
     ,(1, (select user_id from openchpl.user where user_name = 'andlar-test-acb'), null, -2, true)
@@ -59,6 +53,9 @@ insert into openchpl.acl_object_identity (object_id_class, object_id_identity, p
     ,(1, (select user_id from openchpl.user where user_name = 'amore-icsa'), null, -2, true)
     ;
 
+/*
+ * grant acl on each user to themselves
+ */
 insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure) values
     ((select id from openchpl.acl_object_identity where object_id_class = 1 and object_id_identity = (select user_id from openchpl.user where user_name = 'andlar-test')), 0, (select id from openchpl.acl_sid where sid = 'andlar-test'), 16, true, false, false)
     ,((select id from openchpl.acl_object_identity where object_id_class = 1 and object_id_identity = (select user_id from openchpl.user where user_name = 'andlar-test-acb')), 0, (select id from openchpl.acl_sid where sid = 'andlar-test-acb'), 16, true, false, false)
@@ -66,6 +63,11 @@ insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, grant
     ,((select id from openchpl.acl_object_identity where object_id_class = 1 and object_id_identity = (select user_id from openchpl.user where user_name = 'dlucas-icsa')), 0, (select id from openchpl.acl_sid where sid = 'dlucas-icsa'), 16, true, false, false)
     ,((select id from openchpl.acl_object_identity where object_id_class = 1 and object_id_identity = (select user_id from openchpl.user where user_name = 'amore-icsa')), 0, (select id from openchpl.acl_sid where sid = 'amore-icsa'), 16, true, false, false)
     ;
+
+/*
+ * grant acl on acbs for each user that wants them
+ * must have separate insert statement for each specific user/ACB, otherwise MAX function creates bad primary-key relationship
+ */
 insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure) values
     ((select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'Drummond Group')), (select max(ace_order) + 1 from openchpl.acl_entry where acl_object_identity = (select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'Drummond Group'))), (select id from openchpl.acl_sid where sid = 'andlar-test-acb'), 16, true, false, false)
     ,((select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'ICSA Labs')), (select max(ace_order) + 1 from openchpl.acl_entry where acl_object_identity = (select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'ICSA Labs'))), (select id from openchpl.acl_sid where sid = 'andlar-test-acb'), 16, true, false, false)
@@ -80,14 +82,3 @@ insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, grant
 insert into openchpl.acl_entry (acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure) values
     ((select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'ICSA Labs')), (select max(ace_order) + 1 from openchpl.acl_entry where acl_object_identity = (select id from openchpl.acl_object_identity where object_id_class = 2 and object_id_identity = (select certification_body_id from openchpl.certification_body where name = 'ICSA Labs'))), (select id from openchpl.acl_sid where sid = 'amore-icsa'), 16, true, false, false)
     ;
-
-
-    /*
- * When ACLs are changed related to pending certified products, remove from here down
- */
-select openchpl.insert_sid_for_pending_cp((select id from openchpl.acl_sid where sid = 'andlar-test-acb'));
-select openchpl.insert_sid_for_pending_cp((select id from openchpl.acl_sid where sid = 'kekey-icsa'));
-select openchpl.insert_sid_for_pending_cp((select id from openchpl.acl_sid where sid = 'dlucas-icsa'));
-select openchpl.insert_sid_for_pending_cp((select id from openchpl.acl_sid where sid = 'amore-icsa'));
-
-drop function openchpl.insert_sid_for_pending_cp(bigint);